Add has_transition
Function add_transitions checks if transition exists regardless of if conditions are met.

This is similar to ``can_proceed`` without checking conditions are met.

Use case:

Allow to display form button "Complete" even if no all conditions are meet. (When use submit form, there is a chance that conditions are met).